shorelineforecasting/requirements.txt
absl-py==0.9.0
arviz==0.6.1
astor==0.7.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
blinker==1.4
branca==0.4.0
cachetools==3.1.1
certifi==2020.6.20
cffi==1.13.2
cftime==1.0.4.2
chardet==3.0.4
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
convertdate==2.1.3
cryptography==2.8
cycler==0.10.0
Cython==0.29.15
decorator==4.4.1
defusedxml==0.6.0
descartes==1.1.0
entrypoints==0.3
ephem==3.7.7.0
fbprophet==0.6
Fiona==1.8.13
Flask==1.1.1
folium==0.10.1
gast==0.2.2
GDAL==3.0.3
geopandas==0.6.2
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
holidays==0.10.1
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jitcdde==1.4.0
jitcxde-common==1.4.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
LunarCalendar==0.0.9
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mpmath==1.1.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.4
netCDF4==1.5.3
notebook==6.0.3
numpy==1.17.5
oauthlib==3.0.1
opt-einsum==0+untagged.53.g6ab433b.dirty
packaging==20.1
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.1
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.5.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.4
psycopg2==2.8.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
Pygments==2.5.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyproj==2.4.2.post1
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyrsistent==0.15.7
PySocks==1.7.1
pystan==2.19.1.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.23.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==4.0
Rtree==0.9.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
SQLAlchemy==1.3.19
statsmodels==0.10.2
symengine==0.4.0
sympy==1.5.1
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
timesynth==0.2.4
torch==1.4.0
tornado==6.0.3
tqdm==4.45.0
traitlets==4.3.3
urllib3==1.25.7
vincent==0.4.4
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.15.0
zipp==2.2.0
